feat: refactored cjsToEsm.cjs
